feat (shared): système event dispatcher (#80)
