+++ docs/feature.md
@@ -8,7 +8,6 @@ Pour faciliter l'intégration de notre domaine métier, l'utilisation de CQRS es
 
 Actuellement, nous n'utilisons pas du tout le package `symfony/messenger`, car nous souhaitons rester agnostique de ce composant. 
 Son intégration dans l'application consistera à paramétrer les bus et implémenter les adapteurs des différentes interfaces :
-* [EventBus](../src/Shared/Event/EventBus.php)
 * [CommandBus](../src/Shared/Command/CommandBus.php)
 * [QueryBus](../src/Shared/Query/QueryBus.php)
 
@@ -25,7 +24,6 @@ Voici quelques exemples ainsi que des tests associés :
       * **Event** : [NewRegistration](../src/Player/Register/NewRegistration.php)
       * **Handler** : [Register](../src/Player/Register/Register.php)
       * **Command**: [Registration](../src/Player/Register/Registration.php)
-    * **EventListener** : [CreateRegistrationToken](../src/Player/CreateRegistrationToken/CreateRegistrationToken.php) 
     * **Test** : [RegisterTest](../tests/Player/RegisterTest.php)
 * Query 
   * Récupérer un joueur-euse en fonction de ton token d'oubli de mot de passe :
