+++ src/Path/BeginTraining/TrainingBegan.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace IncentiveFactory\Domain\Path\BeginTraining;
 
 use IncentiveFactory\Domain\Path\Path;
-use IncentiveFactory\Domain\Shared\Event\Event;
+use IncentiveFactory\Domain\Shared\EventDispatcher\Event;
 
 final class TrainingBegan implements Event
 {
