+++ src/Player/Register/NewRegistration.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace IncentiveFactory\Domain\Player\Register;
 
 use IncentiveFactory\Domain\Player\Player;
-use IncentiveFactory\Domain\Shared\Event\Event;
+use IncentiveFactory\Domain\Shared\EventDispatcher\Event;
 
 final class NewRegistration implements Event
 {
