+++ src/Player/Register/Register.php
@@ -7,7 +7,7 @@ namespace IncentiveFactory\Domain\Player\Register;
 use IncentiveFactory\Domain\Player\Player;
 use IncentiveFactory\Domain\Player\PlayerGateway;
 use IncentiveFactory\Domain\Shared\Command\CommandHandler;
-use IncentiveFactory\Domain\Shared\Event\EventBus;
+use IncentiveFactory\Domain\Shared\EventDispatcher\EventDispatcher;
 use IncentiveFactory\Domain\Shared\Uid\UlidGeneratorInterface;
 use IncentiveFactory\Domain\Shared\Uid\UuidGeneratorInterface;
 use Symfony\Component\PasswordHasher\PasswordHasherInterface;
@@ -19,7 +19,7 @@ final class Register implements CommandHandler
         private UlidGeneratorInterface $ulidGenerator,
         private UuidGeneratorInterface $uuidGenerator,
         private PlayerGateway $playerGateway,
-        private EventBus $eventBus
+        private EventDispatcher $eventBus
     ) {
     }
