--- src/Shared/Event/Event.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace IncentiveFactory\Domain\Shared\Event;
-
-interface Event
-{
-}
