+++ src/Shared/EventDispatcher/Event.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace IncentiveFactory\Domain\Shared\EventDispatcher;
+
+interface Event
+{
+}
