+++ src/Shared/EventDispatcher/EventDispatcher.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace IncentiveFactory\Domain\Shared\Event;
+namespace IncentiveFactory\Domain\Shared\EventDispatcher;
 
-interface EventBus
+interface EventDispatcher
 {
     public function dispatch(Event $event): void;
 }
